Clock Tag Registry

Lists the keyword tags that expand to a group of clocks, for use as a clocks or tag value.

Usage

list_clock_tags()

Value

A named list. Each name is a tag, and each element is a character vector of the group or clock tokens it expands to.

See also

Examples

list_clock_tags()
#> $gestational
#> [1] "Bohlin"          "Knight"          "Mayne"           "LeePlacentalAge"
#> 
#> $mitotic
#> [1] "EpiTOC"    "EpiTOC2"   "MiAge"     "RepliTali"
#> 
#> $mortality
#> [1] "GrimAge"        "ZhangMortality"
#>